feat(#13): Create addressgenerator class with db connection to get fire hydrant addresss by lane and intersection
